Sprint2

See merge request idt-lw/m8o-103bv-25/personal/dokukarin/cacheflow!14

+++ frontend/shared/interop-sample/domain/src/commonMain/kotlin/interopSample/models/Weather.kt
@@ -1,7 +1,10 @@
 package interopSample.models
 
+import utils.types.BigDecimal
+import kotlin.js.JsExport
+
 @JsExport
 data class Weather(
-    val temperature: Double,
+    val temperature: BigDecimal,
     val temperatureUnit: String
 )
